pyproject.toml
[project]
name = "hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==3.0.0",
    "apache-airflow-providers-amazon==9.0.0",
    "apache-airflow-providers-postgres>=6.2.0",
    "notebook>=7.4.4",
    "xmltodict>=0.14.2",
]
